code refactoring

+++ server.py
@@ -133,6 +133,8 @@ def get_user(user_id):
 @app.route("/user/<user_id>", methods=["DELETE"])
 def delete_user(user_id):
     user = User.query.filter_by(id=user_id).first()
+    if not user:
+        return jsonify({"message": "user does not exist"}), 400
     db.session.delete(user)
     db.session.commit()
     return jsonify({"message": "user deleted"}), 200
